backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Utilities
pytz==2024.1
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
faker==22.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
